requirements.txt
easyocr==1.7.2
Flask==3.1.0
numpy==2.1.3
Pillow==11.0.0
